+++ src/test/java/io/github/astrapi69/io/file/filter/PrefixFileFilterTest.java
@@ -60,7 +60,7 @@ public class PrefixFileFilterTest
 
 		filename = "resources.properties";
 
-		filepath = "de/alpharogroup/lang/" + filename;
+		filepath = "io/github/astrapi69/lang/" + filename;
 
 		file = ClassExtensions.getResourceAsFile(filepath);
 		dir = file.getParentFile();