+++ src/test/java/io/github/astrapi69/io/file/namefilter/MultiplyExtensionsFilenameFilterTest.java
@@ -78,7 +78,7 @@ public class MultiplyExtensionsFilenameFilterTest
 
 		filename = "resources.properties";
 
-		propertiesFilename = "de/alpharogroup/lang/" + filename;
+		propertiesFilename = "io/github/astrapi69/lang/" + filename;
 
 		file = ClassExtensions.getResourceAsFile(propertiesFilename);
 		dir = file.getParentFile();